requirements.txt
pyrogram
tgcrypto
py-tgcalls==0.5.5
python-dotenv
youtube_dl
youtube_search_python==1.4.6
googletrans==4.0.0rc1
gTTS
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg-python
pillow
future
psutil
wget
motor
dnspython
bs4
pyaml
yt_dlp
httpx==0.13.3
pathlib